MNP-22: add migration on running
