+++ service/.dockerignore
@@ -1,4 +1,3 @@
-node_modules
 npm-debug.log
 yarn-error.log
 pnpm-lock.yaml
