+++ service/docker/dev.docker-compose.yml
@@ -12,6 +12,7 @@ services:
     environment:
       NODE_ENV: development
       PORT: 8080
+      WITH_MIGRATION:
     volumes:
       - ../:/app
       - /app/node_modules
