+++ service/src/index.ts
@@ -1,10 +1,17 @@
 import { getHttpServer } from './http/getHttpServer';
 import { getWebSocketServer } from './wss/getWebSocketServer';
 import { gracefulShutdown } from './gracefulShutdown';
-import { getConfig } from './getConfig';
+import { getConfig } from './nodejs/getConfig';
+import { migrate } from './migrations/migration';
+import { getConnectedPostgresClient } from './getConnectedPostgresClient';
 
 const main = async () => {
     const config = getConfig();
+    const client = await getConnectedPostgresClient(config.postgresConfig);
+
+    if (config.withMigration) {
+        await migrate(client);
+    }
 
     const httpServer = getHttpServer();
     const wss = getWebSocketServer();
