+++ service/src/nodejs/toBoolean.ts
@@ -1,0 +1,3 @@
+export const toBoolean = (value?: string): boolean => {
+  return value === 'true';
+};
